requirements_enhanced.txt
Flask
line-bot-sdk
requests
gunicorn
google-cloud-firestore
python-dotenv
aiohttp
asyncio
openai
google-cloud-dialogflow
